User management logic moved to separate file
