+++ app/__init__.py
@@ -46,8 +46,10 @@ def create_app():
              use_ssl=False,
              base_url=None)
 
-    from .main import auth, main
-    app.register_blueprint(auth)
+    from app.main import main
     app.register_blueprint(main)
+
+    from app.auth import auth
+    app.register_blueprint(auth)
 
     return app
