+++ app/main.py
@@ -1,18 +1,19 @@
 from datetime import date
-from flask import Blueprint, render_template, request, redirect, url_for, flash
 
-from flask_login import login_user, login_required, current_user, logout_user
-from app.forms import CreatePostForm, LoginForm, RegisterForm, CommentForm
-from sqlalchemy import exc
-from werkzeug.security import generate_password_hash, check_password_hash
-from app.constants import DB_PATH
-from app import db
-from app.models import *
-import os
-from app import create_app
+from flask import (
+    Blueprint,
+    render_template,
+    request,
+    redirect,
+    url_for,
+    flash
+)
+from flask_login import current_user
+
 from app.decorators import admin_only
+from app.forms import CreatePostForm, CommentForm
+from app.models import *
 
-auth = Blueprint('auth', __name__)
 main = Blueprint('main', __name__)
 
 
@@ -20,49 +21,6 @@ main = Blueprint('main', __name__)
 def get_all_posts():
     posts = BlogPost.query.all()
     return render_template("index.html", all_posts=posts)
-
-
-@auth.route('/register', methods=["GET", "POST"])
-def register():
-    form = RegisterForm()
-    if form.validate_on_submit():
-        try:
-            new_user = User(email=request.form["email"], password=generate_password_hash(password=request.form["password"], salt_length=8), name=request.form["name"])
-            db.session.add(new_user)
-            db.session.commit()
-            login_user(new_user)
-            return redirect(url_for("main.get_all_posts"))
-        except exc.IntegrityError:
-            flash("User already exist. Please log in.")
-            return redirect(url_for(".login"))
-    else:
-        return render_template("register.html", form=form)
-
-
-@auth.route('/login', methods=["GET", "POST"])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=request.form["email"]).first()
-        if user:
-            if check_password_hash(user.password, request.form["password"]):
-                login_user(user)
-                return redirect(url_for("main.get_all_posts"))
-            else:
-                flash("Wrong password")
-                return redirect(url_for(".login"))
-        else:
-            flash("This email is not registered")
-            return redirect(url_for(".login"))
-    else:
-        return render_template("login.html", form=form)
-
-
-@auth.route('/logout')
-@login_required
-def logout():
-    logout_user()
-    return redirect(url_for('main.get_all_posts'))
 
 
 @main.route("/post/<int:post_id>", methods=["GET", "POST"])
