+++ run.py
@@ -14,4 +14,4 @@ if __name__ == "__main__":
     app = create_app()
     create_db(app)
 
-    app.run()
+    app.run(host="127.0.0.1", port=5000)
